Split owned liquidity definitions from price handlers
