+++ subgraphs/berachain/src/price/PriceLookup.ts
@@ -1,34 +1,10 @@
 import { Address, BigDecimal, BigInt, log } from "@graphprotocol/graph-ts";
 
-import { PriceHandler, PriceLookupResult } from "../../../shared/src/price/PriceHandler";
-import { PriceHandlerRemapping } from "../../../shared/src/price/PriceHandlerRemapping";
-import { PriceHandlerUniswapV3Quoter } from "../../../shared/src/price/PriceHandlerUniswapV3Quoter";
+import { PriceLookupResult } from "../../../shared/src/price/PriceHandler";
 import { getUSDRate } from "../../../shared/src/price/PriceRouter";
-import {
-  BERADROME_KODIAK_OHM_HONEY_REWARD_VAULT,
-  ERC20_HONEY,
-  ERC20_IBERA,
-  ERC20_OHM,
-  ERC20_WBERA,
-  LP_BERADROME_KODIAK_OHM_HONEY,
-  LP_KODIAK_IBERA_WBERA,
-  LP_KODIAK_OHM_HONEY,
-  LP_UNISWAP_V3_WBERA_HONEY,
-  NATIVE_BERA} from "../contracts/Constants";
 import { getContractName } from "../contracts/Contracts";
+import { PRICE_HANDLERS } from "../contracts/LiquidityConstants";
 import { getBaseTokenRate, isBaseToken } from "./PriceBase";
-import { PriceHandlerKodiakIsland } from "./PriceHandlerKodiakIsland";
-
-const KODIAK_QUOTER = "0x644C8D6E501f7C994B74F5ceA96abe65d0BA662B".toLowerCase();
-
-// Price lookup is enabled for these tokens
-export const PRICE_HANDLERS: PriceHandler[] = [
-  new PriceHandlerUniswapV3Quoter([ERC20_HONEY, ERC20_WBERA], KODIAK_QUOTER, LP_UNISWAP_V3_WBERA_HONEY, getContractName),
-  new PriceHandlerKodiakIsland([ERC20_HONEY, ERC20_OHM], KODIAK_QUOTER, LP_KODIAK_OHM_HONEY, null, null, getContractName),
-  new PriceHandlerKodiakIsland([ERC20_HONEY, ERC20_OHM], KODIAK_QUOTER, LP_KODIAK_OHM_HONEY, BERADROME_KODIAK_OHM_HONEY_REWARD_VAULT, LP_BERADROME_KODIAK_OHM_HONEY, getContractName),
-  new PriceHandlerUniswapV3Quoter([ERC20_IBERA, ERC20_WBERA], KODIAK_QUOTER, LP_KODIAK_IBERA_WBERA, getContractName),
-  new PriceHandlerRemapping(NATIVE_BERA, ERC20_WBERA, getContractName),
-];
 
 /**
  * Internal function to determine the price, using the shared price functions.
